% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gam_formula.R
\name{build_gam_knots}
\alias{build_gam_knots}
\title{Build the knots list for gam() from earth knot data}
\usage{
build_gam_knots(smooth_specs, earth_knots)
}
\arguments{
\item{smooth_specs}{List of smooth-term specs (same as
\code{\link[=build_gam_formula]{build_gam_formula()}}).}

\item{earth_knots}{An \code{mgcvUI_earth_knots} object, or \code{NULL}.}
}
\value{
A named list suitable for the \code{knots} argument of
\code{\link[mgcv:gam]{mgcv::gam()}}, or \code{NULL} if no knots apply.
}
\description{
When earth knots are available for variables using \code{"cr"} or \code{"ps"}
basis types, this function prepares the \code{knots} argument for
\code{\link[mgcv:gam]{mgcv::gam()}}.
}
\examples{
specs <- list(
  list(vars = "wt", type = "s", bs = "cr", k = 5L)
)
build_gam_knots(specs, NULL)
}
