% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regproj.R
\name{mgcvui_prefs_read}
\alias{mgcvui_prefs_read}
\title{Read user preferences (returns empty list if file missing)}
\usage{
mgcvui_prefs_read()
}
\value{
Named list.
}
\description{
Read user preferences (returns empty list if file missing)
}
