% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model.R
\name{mod_model_fit_ui}
\alias{mod_model_fit_ui}
\title{Model Fit Button — UI (sidebar)}
\usage{
mod_model_fit_ui(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A \link[shiny:reexports]{shiny::tagList}.
}
\description{
Model Fit Button — UI (sidebar)
}
\examples{
if (interactive()) {
  ui <- fluidPage(mod_model_fit_ui("model1"))
}
}
