% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model.R
\name{mod_model_server}
\alias{mod_model_server}
\title{Model Module — Server (shared by fit button and results)}
\usage{
mod_model_server(id, data_r, var_config_r, earth_knots_r = reactive(NULL))
}
\arguments{
\item{id}{Shiny module namespace ID.}

\item{data_r}{A reactive returning the data frame.}

\item{var_config_r}{A reactive returning the variable configuration.}

\item{earth_knots_r}{A reactive returning earth knots (or \code{NULL}).}
}
\value{
A reactive containing the \code{mgcvUI_result}, or \code{NULL}.
}
\description{
Model Module — Server (shared by fit button and results)
}
