% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_actual_vs_predicted}
\alias{plot_actual_vs_predicted}
\title{Plot Actual vs Predicted}
\usage{
plot_actual_vs_predicted(gam_result)
}
\arguments{
\item{gam_result}{An \code{mgcvUI_result} from \code{\link[=fit_gam]{fit_gam()}}.}
}
\value{
A ggplot object.
}
\description{
Plot Actual vs Predicted
}
\examples{
specs <- list(list(vars = "wt", type = "s", bs = "tp", k = NULL))
res <- fit_gam(mtcars, "mpg", specs)
plot_actual_vs_predicted(res)
}
