% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_by_smooth_interactive}
\alias{plot_by_smooth_interactive}
\title{Interactive Factor-by-Smooth Contribution Plot}
\usage{
plot_by_smooth_interactive(gam_result, variable, by_var, dark_mode = FALSE)
}
\arguments{
\item{gam_result}{An \code{mgcvUI_result} from \code{\link[=fit_gam]{fit_gam()}}.}

\item{variable}{Character name of the continuous variable.}

\item{by_var}{Character name of the factor variable.}

\item{dark_mode}{Logical; use dark theme styling. Default \code{FALSE}.}
}
\value{
A plotly htmlwidget.
}
\description{
Creates a plotly line plot for a factor-by-smooth term,
showing one contribution curve per factor level.
}
