% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilogy.R
\name{trilogy_is_locked}
\alias{trilogy_is_locked}
\title{Is the earth model locked for this project?}
\usage{
trilogy_is_locked(project_path)
}
\arguments{
\item{project_path}{Absolute path to the regProj appraisal project (the flat
segment folder under \verb{<root>/appr/}).}
}
\value{
Logical scalar.
}
\description{
Is the earth model locked for this project?
}
