% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_server}
\alias{mod_data_server}
\title{Data Import Module — Server}
\usage{
mod_data_server(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A list with \code{data} (reactive data frame) and \code{filename}
(reactive original file basename).
}
\description{
Data Import Module — Server
}
