% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_ui}
\alias{mod_data_ui}
\title{Data Import Module -- UI}
\usage{
mod_data_ui(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A \link[shiny:reexports]{shiny::tagList}.
}
\description{
Compact file upload widget for embedding in a sidebar.
}
\examples{
if (interactive()) {
  ui <- fluidPage(mod_data_ui("data1"))
  server <- function(input, output, session) {
    mod_data_server("data1")
  }
  shinyApp(ui, server)
}
}
