% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_earth_import.R
\name{mod_earth_import_server}
\alias{mod_earth_import_server}
\title{Earth Import Module -- Server}
\usage{
mod_earth_import_server(id)
}
\arguments{
\item{id}{Shiny module namespace ID.}
}
\value{
A reactive containing an \code{mgcvUI_earth_knots} object, or
\code{NULL}.
}
\description{
Earth Import Module -- Server
}
