% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_variables.R
\name{mod_variables_server}
\alias{mod_variables_server}
\title{Variable Selection Module -- Server}
\usage{
mod_variables_server(
  id,
  data_r,
  filename_r = reactive(NULL),
  earth_knots_r = reactive(NULL)
)
}
\arguments{
\item{id}{Shiny module namespace ID.}

\item{data_r}{A reactive returning the imported data frame.}

\item{filename_r}{A reactive returning the original filename.}

\item{earth_knots_r}{A reactive returning an \code{mgcvUI_earth_knots}
object (or \code{NULL}).}
}
\value{
A reactive list with components: \code{response}, \code{predictors},
\code{smooth_specs}, \code{family}, \code{method}, \code{select}, \code{gamma}.
}
\description{
Variable Selection Module -- Server
}
